Improving transaction system and error recovery
A new state is written as a placeholder until all servers have
been checked for changes.
This makes sure we don't add items and remove them again in the
same sync.

In case of an error when writing the changes to the server,
the todo is kept open and written back to the state.json

In case we have an open todo for a particular item on a certain
server, we ignore certain actions until all open todos have been
resolved on the server.
This should make sure we don't loose data, instead we might
restore already deleted items because of that, but that shouldn't
be a big problem.

After updating a playlist, make sure the playlist was really updated.
Sometimes the server refuses to update a playlist because of connection
problems to YT.
We need to keep the todo in that case and try it again later.